bfa/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
a2a-sdk
rank-bm25
numpy